Changes for local installation on Docker.
Former-commit-id: 193b54c0b2ebab3469c4a7e9200d4dd91ac2d92e [formerly 315fc9e7e7c52c5306fd4a58d30efd3801cb9ed7] [formerly c5cf5d5028ef9075c6195f1177ffa35dce9e5a5b [formerly e50d20040f3de7c36a8884a9970e823b89e08383]] [formerly 332bbbeb84cb2b44da507663972fbd6517fc61d7] [formerly 621b90e2dfd6b4cab2a7f67aafeff85862319f4b] [formerly 7601c374facebf9b484fb2c6da2c47e35dbd25bf [formerly e69ac1f96c7e23694cba8b714f3c5a7f84f1014f] [formerly 621b90e2dfd6b4cab2a7f67aafeff85862319f4b [formerly 621b90e2dfd6b4cab2a7f67aafeff85862319f4b [formerly 621b90e2dfd6b4cab2a7f67aafeff85862319f4b [formerly 104690ef9c162fbc2d7c8a806ecc0bd7cc63583b]]]]]
Former-commit-id: 78d7e3475e691873f937488554d824a285c95e7f [formerly d29645453a235c04099587bbf62d00011082e0e7] [formerly 26a6d1c812abfff326d89c6d7f127cb9a139ccdc [formerly 26a6d1c812abfff326d89c6d7f127cb9a139ccdc [formerly 26a6d1c812abfff326d89c6d7f127cb9a139ccdc [formerly 75496e4295324e2bf05918bca1556d3e1da541a0]]]]
Former-commit-id: b8d3631654661d965f859e143036d30ce08de9ba [formerly 26a6d1c812abfff326d89c6d7f127cb9a139ccdc]
Former-commit-id: 8fca4e8effeefa56c5ec4abeac4f0aa51493fe83
Former-commit-id: 56db61b282b6219f582213a268ca304b64eae209 [formerly fa77615f9a11e327cfa9fe981a46c769fee8f185]
Former-commit-id: 6c0d2c60e2345a09fd85b88d3b1a36d0b885547c

+++ scripts/elk-setup/agent/setup-centos.sh
@@ -8,12 +8,23 @@ logstashForwarderRepoFilePath=$logstashForwarderConfigDirectoryPath/logstash-for
 logstashForwarderInstallablesDirectoryPath=$2
 
 gpgKeyElasticsearchFilePath=$logstashForwarderInstallablesDirectoryPath/GPG-KEY-elasticsearch
-logstashForwarderRPMFilePath=$logstashForwarderInstallablesDirectoryPath/check_mk-agent-1.2.4p5-1.noarch.rpm
+logstashForwarderRPMFilePath=$logstashForwarderInstallablesDirectoryPath/logstash-forwarder-0.4.0-1.x86_64.rpm
 
 echo ...........................Installing logstash-forwarder...........................
-rpm --import http://packages.elasticsearch.org/GPG-KEY-elasticsearch
+
+if [ -f $gpgKeyElasticsearchFilePath ]; then
+	rpm --import $gpgKeyElasticsearchFilePath
+else
+	rpm --import http://packages.elasticsearch.org/GPG-KEY-elasticsearch
+fi
+
 cp $logstashForwarderRepoFilePath  /etc/yum.repos.d
-yum -y install logstash-forwarder
+
+if [ -f $logstashForwarderRPMFilePath ]; then
+	yum install -y $logstashForwarderRPMFilePath
+else
+	yum -y install logstash-forwarder
+fi
 
 cp $logstashForwarderCertificateFilePath /etc/pki/tls/certs
 cp $logstashForwarderConfigFilePath  /etc
